New upstream version 1.4.0~beta

+++ src/knot/server/journal.c
@@ -25,9 +25,9 @@
 #include <sys/mman.h>
 
 #include "common/crc.h"
-#include "common.h"
+#include "libknot/common.h"
 #include "knot/other/debug.h"
-#include "journal.h"
+#include "knot/server/journal.h"
 
 /*! \brief Infinite file size limit. */
 #define FSLIMIT_INF (~((size_t)0))
@@ -644,7 +644,7 @@ journal_t* journal_open(const char *fn, size_t fslimit, int mode, uint16_t bflag
 	if (fslimit == 0) {
 		j->fslimit = FSLIMIT_INF;
 	} else {
-		j->fslimit = (size_t)fslimit;
+		j->fslimit = fslimit;
 	}
 
 	dbg_journal("journal: opened journal size=%u, queue=<%u, %u>, fd=%d\n",
